:art: split Decoder.swift into several extensions
